Commit after ignoring for ages
Fix Client.createParcel typo
TrackingEvents are not check for datetime uniqueness if datetime is -1.
This change is shit, but justified by setting datetime to -1 before
merging so dateTime can reflect either when we saw the event, or use
the self-reported event time.
....
Just add a "firstSeen" column to the events that isn't compared. Duh.
Debug log queue controller
Add shopify data types